fix(cli): address US1 review findings
Based on comprehensive code review , this commit
addresses all identified issues with the US1 implementation:

MUST FIX Issues Resolved:
1. Documentation sync: Updated README.md and docs/usage.md to reflect
   default max_retries=5 (was incorrectly documented as 3)

2. Global flag consistency: Wire --max-retries through sources command
   - sources.rs now accepts max_retries parameter
   - main.rs passes cli.max_retries to sources.execute()
   - Fixes inconsistency where global flag was ignored by sources list

SHOULD FIX Issues Resolved:
3. Retry message semantics: Change denominator from (max_retries+1) to max_retries
   - binance_http.rs: max_attempts = self.max_retries (not +1)
   - With --max-retries 10, logs now show "attempt X/10" not "attempt X/11"
   - Aligns with user expectation and spec scenario 3

4. Type safety improvements: CLI now uses u32 with range validation
   - download.rs: max_retries changed from usize to u32
   - Added clap value_parser with range 1..=20
   - Eliminates unsafe narrowing conversion (usize -> u32)
   - Prevents unrealistic values (>20 retries)

5. Removed unnecessary type conversions:
   - main.rs, download.rs: cli.max_retries (no longer 'as u32')
   - Type matches throughout chain: u32 end-to-end

All tests passing (47 lib + 5 integration tests for US1)

Review verdict: APPROVE WITH CHANGES (all changes implemented)

+++ src/cli/download.rs
@@ -68,9 +68,9 @@ pub struct Cli {
     #[arg(long, global = true, default_value = "1")]
     pub concurrency: usize,
 
-    /// Maximum number of retries for failed requests (default: 5)
-    #[arg(long, global = true, default_value = "5")]
-    pub max_retries: usize,
+    /// Maximum number of retries for failed requests (default: 5, range: 1-20)
+    #[arg(long, global = true, default_value = "5", value_parser = clap::value_parser!(u32).range(1..=20))]
+    pub max_retries: u32,
 
     /// Force re-download even if output file already covers the requested range
     #[arg(long, global = true, default_value_t = false)]
@@ -424,7 +424,7 @@ impl BarsArgs {
 
         // Apply CLI flags to executor (P0-6 fix)
         let executor = executor
-            .with_max_retries(cli.max_retries as u32)
+            .with_max_retries(cli.max_retries)
             .with_force(cli.force)
             .with_shutdown(shutdown);
 
@@ -553,7 +553,7 @@ impl AggTradesArgs {
 
         // Apply CLI flags to executor (P0-6 fix)
         let executor = executor
-            .with_max_retries(cli.max_retries as u32)
+            .with_max_retries(cli.max_retries)
             .with_force(cli.force)
             .with_shutdown(shutdown);
 
@@ -842,7 +842,7 @@ impl FundingArgs {
 
         // Apply CLI flags to executor (P0-6 fix)
         let executor = executor
-            .with_max_retries(cli.max_retries as u32)
+            .with_max_retries(cli.max_retries)
             .with_force(cli.force)
             .with_shutdown(shutdown);
